requirements.txt
aioapns==3.3.1
aiohttp==3.8.5
aiosignal==1.4.0
amqp==5.3.1
asgiref==3.11.1
async-timeout==4.0.2
attrs==25.4.0
beautifulsoup4==4.14.3
billiard==4.2.4
bs4==0.0.2
celery==5.6.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coreapi==2.3.3
drf-spectacular==0.27.2
drf-spectacular[sidecar]==0.27.2
aioapns==3.3.1
aiohttp==3.8.5
aiosignal==1.4.0
amqp==5.3.1
asgiref==3.11.1
async-timeout==4.0.2
attrs==25.4.0
beautifulsoup4==4.14.3
billiard==4.2.4
bs4==0.0.2
celery==5.6.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
cron_descriptor==2.0.6
cryptography==46.0.4
diff-match-patch==20241021
Django==4.2.27
django-admin-interface==0.30.1
django-allauth==65.11.1
django-celery-beat==2.8.1
django-cogwheels==0.3
django-colorfield==0.14.0
django-cors-headers==4.8.0
django-crispy-forms==2.4
django-environ==0.12.0
django-filter==24.3
django-hijack==3.7.4
django-import-export==4.3.9
django-jazzmin==3.0.1
django-jet-reboot==1.3.10
django-modelcluster==6.4
django-permissionedforms==0.1
django-redis==6.0.0
django-stubs-ext==5.2.5
django-taggit==6.1.0
django-tasks==0.8.1
django-timezone-field==7.1
django-treebeard==4.7.1
django-walletpass==4.2.1
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dotenv==0.9.9
frozenlist==1.8.0
h2==4.3.0
hpack==4.1.0
hyperframe==6.1.0
idna==3.11
itypes==1.2.0
Jinja2==3.1.6
kombu==5.6.2
MarkupSafe==3.0.3

# Required for DRF docs
coreapi==2.3.3
multidict==6.7.1
packaging==26.0
pillow==12.1.0
prompt_toolkit==3.0.52
propcache==0.4.1
psycopg==3.3.2
psycopg-binary==3.3.2
pycparser==3.0
PyJWT==2.11.0
pyOpenSSL==25.3.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-slugify==8.0.4
redis==7.1.0
requests==2.32.5
six==1.17.0
soupsieve==2.8.3
sqlparse==0.5.5
tablib==3.9.0
text-unidecode==1.3
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
yarl==1.22.0